# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Auction Hub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  both: "both",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  rating: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
  both: "both",
} as const;

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  auctionCount: number;
}

export type AuctionStatus = (typeof AuctionStatus)[keyof typeof AuctionStatus];

export const AuctionStatus = {
  pending: "pending",
  live: "live",
  ended: "ended",
  sold: "sold",
} as const;

export interface Auction {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  category: string;
  startingPrice: number;
  currentBid: number;
  bidCount: number;
  sellerId: number;
  sellerName: string;
  sellerRating: number;
  status: AuctionStatus;
  startTime: string;
  endTime: string;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerName?: string | null;
  /** @nullable */
  shippingInfo?: string | null;
  createdAt: string;
}

export interface Bid {
  id: number;
  auctionId: number;
  bidderId: number;
  bidderName: string;
  amount: number;
  createdAt: string;
}

export type AuctionDetail = Auction & {
  recentBids: Bid[];
};

export interface SoldItem {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  category: string;
  soldAmount: number;
  /** @nullable */
  winnerName?: string | null;
  /** @nullable */
  winnerId?: number | null;
  soldAt: string;
  /** @nullable */
  shippingInfo?: string | null;
}

export interface CreateAuctionBody {
  title: string;
  description: string;
  imageUrl: string;
  category: string;
  startingPrice: number;
  startTime: string;
  endTime: string;
  shippingInfo?: string;
}

export type UpdateAuctionBodyStatus =
  (typeof UpdateAuctionBodyStatus)[keyof typeof UpdateAuctionBodyStatus];

export const UpdateAuctionBodyStatus = {
  pending: "pending",
  live: "live",
  ended: "ended",
  sold: "sold",
} as const;

export interface UpdateAuctionBody {
  status?: UpdateAuctionBodyStatus;
  endTime?: string;
}

export interface AcceptWinnerBody {
  winnerId: number;
}

export type BidWithAuction = Bid & {
  auctionTitle: string;
  auctionStatus: string;
  auctionEndTime: string;
  isWinning: boolean;
};

export interface PlaceBidBody {
  amount: number;
}

export interface SellerStats {
  totalAuctions: number;
  activeAuctions: number;
  totalEarnings: number;
  pendingAuctions: number;
  soldAuctions: number;
  totalBidsReceived: number;
}

export interface AuctionsSummary {
  liveCount: number;
  totalBids: number;
  totalValue: number;
  endingSoon: number;
}

export interface ActivityItem {
  id: number;
  bidderName: string;
  auctionTitle: string;
  auctionId: number;
  amount: number;
  createdAt: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListAuctionsParams = {
  category?: string;
  status?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sort?: string;
};