# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Auction Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auctions:
    get:
      operationId: listAuctions
      tags: [auctions]
      summary: List auctions with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of auctions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Auction"

    post:
      operationId: createAuction
      tags: [auctions]
      summary: Create a new auction (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuctionBody"
      responses:
        "201":
          description: Auction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auctions/{id}:
    get:
      operationId: getAuction
      tags: [auctions]
      summary: Get auction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Auction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuctionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAuction
      tags: [auctions]
      summary: Update auction (seller only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuctionBody"
      responses:
        "200":
          description: Auction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
  /auctions/{id}/accept-winner:
    post:
      operationId: acceptWinner
      tags: [auctions]
      summary: Seller accepts highest bidder as winner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptWinnerBody"
      responses:
        "200":
          description: Winner accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"

  # BIDS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Auction:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        startingPrice:
          type: number
        currentBid:
          type: number
        bidCount:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        sellerRating:
          type: number
        status:
          type: string
          enum: [pending, live, ended, sold]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        shippingInfo:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - imageUrl
        - category
        - startingPrice
        - currentBid
        - bidCount
        - sellerId
        - sellerName
        - sellerRating
        - status
        - startTime
        - endTime
        - createdAt
    AuctionDetail:
      allOf:
        - $ref: "#/components/schemas/Auction"
        - type: object
          properties:
            recentBids:
              type: array
              items:
                $ref: "#/components/schemas/Bid"
          required: [recentBids]
...
    CreateAuctionBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        startingPrice:
          type: number
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        shippingInfo:
          type: string
      required: [title, description, imageUrl, category, startingPrice, startTime, endTime]
    UpdateAuctionBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, live, ended, sold]
        endTime:
          type: string
          format: date-time
    AcceptWinnerBody:
      type: object
      properties:
        winnerId:
          type: integer
      required: [winnerId]
    Bid:
      type: object
      properties:
        id:
          type: integer
        auctionId:
          type: integer
        bidderId:
          type: integer
        bidderName:
          type: string
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, auctionId, bidderId, bidderName, amount, createdAt]
...